pyproject.toml
[tool.poetry]
name = "orodruin"
version = "0.1.0"
description = "A Python rigging graph library."
license = "LICENSE.md"
authors = ["Loïc Pinsard <muream@gmail.com>", "Aphosis <aphosis.github@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/HolisticCoders/orodruin"
repository = "https://github.com/HolisticCoders/orodruin"

[tool.poetry.dependencies]
python = "^3.7"
typing-extensions = "^3.7.4"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
black = "^20.8b1"
pylint = "^2.6.0"
isort = "^5.6.4"
pytest-cov = "^2.10.1"
mypy = "^0.812"
snakeviz = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 88
disable = [
    "duplicate-code",
    "fixme",
    "import-error",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes",
    "invalid-name",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--doctest-modules --cov=orodruin --cov-report html"
testpaths = [
    "tests",
    "orodruin",
]
